+++ kedro_kubeflow/context_helper.py
@@ -5,14 +5,17 @@ from kedro import __version__ as kedro_version
 from semver import VersionInfo
 
 from .config import PluginConfig
-from .kfpclient import KubeflowClient
 
 
 class ContextHelper(object):
 
     CONFIG_FILE_PATTERN = "kubeflow*"
 
-    def __init__(self, metadata, env):
+    def __init__(
+        self,
+        metadata,
+        env,
+    ):
         self._metadata = metadata
         self._env = env
 
@@ -21,12 +24,15 @@ class ContextHelper(object):
         return self._metadata.project_name
 
     @property
-    def context(self):
+    @lru_cache()
+    def session(self):
         from kedro.framework.session import KedroSession
 
-        return KedroSession.create(
-            self._metadata.package_name, env=self._env
-        ).load_context()
+        return KedroSession.create(self._metadata.package_name, env=self._env)
+
+    @property
+    def context(self):
+        return self.session.load_context()
 
     @property
     @lru_cache()
@@ -37,15 +43,37 @@ class ContextHelper(object):
     @property
     @lru_cache()
     def kfp_client(self):
-        return KubeflowClient(self.config, self.project_name, self.context)
+        if self.config.is_vertex_ai_pipelines:
+            from .vertex_ai.client import VertexAIPipelinesClient
+
+            return VertexAIPipelinesClient(
+                self.config, self.project_name, self.context
+            )
+        else:
+            from .kfpclient import KubeflowClient
+
+            return KubeflowClient(
+                self.config,
+                self.project_name,
+                self.context,
+            )
 
     @staticmethod
-    def init(metadata, env):
+    def init(
+        metadata,
+        env,
+    ):
         version = VersionInfo.parse(kedro_version)
         if version.match(">=0.17.0"):
-            return ContextHelper(metadata, env)
+            return ContextHelper(
+                metadata,
+                env,
+            )
         else:
-            return ContextHelper16(metadata, env)
+            return ContextHelper16(
+                metadata,
+                env,
+            )
 
 
 class ContextHelper16(ContextHelper):